move ./error.js to ./misc/error.js
