+++ packages/backend/src/error.ts
@@ -1,12 +1,6 @@
 /*
- * SPDX-FileCopyrightText: MomentQYC and other misskey contributors
+ * SPDX-FileCopyrightText: tamaina and other misskey contributors
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-export function ErrorHandling(message: string): Error {
-	const error = new Error(message);
-	if (process.env.NODE_ENV === 'production') {
-		error.stack = undefined;
-	}
-	return error;
-}
+// Never use `./error.js` because jest can't use it.
