+++ packages/backend/src/server/api/endpoints/reset-password.ts
@@ -8,7 +8,7 @@ import { Inject, Injectable } from '@nestjs/common';
 import type { UserProfilesRepository, PasswordResetRequestsRepository } from '@/models/index.js';
 import { Endpoint } from '@/server/api/endpoint-base.js';
 import { DI } from '@/di-symbols.js';
-import { ErrorHandling } from '@/error.js'; // TODO Line 51
+import { ErrorHandling } from '@/misc/error.js'; // TODO Line 51
 
 export const meta = {
 	tags: ['reset password'],
